Rename edit and save
